fix Windows build (STATUS_XXX conflicts with system headers)

+++ src/history/test/HistoryTests.cpp
@@ -656,7 +656,7 @@ TEST_CASE("Publish catchup alternation with stall",
     REQUIRE(catchupSimulation.catchupOnline(completeApp, checkpointLedger, 5));
     REQUIRE(catchupSimulation.catchupOnline(minimalApp, checkpointLedger, 5));
 
-    for (size_t i = 1; i < 4; ++i)
+    for (int i = 1; i < 4; ++i)
     {
         // Now alternate between publishing new stuff and catching up to it.
         checkpoint += i;
